test orphaned template
